[TRACK_F] FIX: restore `clippy --all-targets -D warnings` to clean on both workspaces
Found while running the documented baseline from NEXT_SESSION.md STEP 3. Four
lint errors were failing that gate on a clean tree; confirmed PRE-EXISTING by
stashing this session's changes and re-running against untouched main, so they
predate this work and were not introduced here.

The narrower `cargo clippy -- -D warnings` in CLAUDE.md does lint test targets,
which is why the gate read as clean while `--all-targets` did not.

  - `riinac/tests/fuzz_robustness.rs`: `(0x20 + rng.below(0x5f) as u8) as u8` —
    the outer cast is u8 -> u8. Dropped it; the value is unchanged, so the fuzz
    corpus is byte-identical.
  - `riina-core/src/crypto/mod.rs`: 3x `assert_eq!(<bool expr>, true/false, ..)`
    -> `assert!` / `assert!(!..)`, messages preserved. HMAC verify-accepts and
    two verify-rejects assertions keep exactly their previous meaning.

Verified: `cargo clippy --all-targets -- -D warnings` exits 0 on BOTH 03_PROTO
and 05_TOOLING; `cargo test --all --manifest-path 05_TOOLING/Cargo.toml` = 304
pass / 0 fail.

+++ 03_PROTO/crates/riinac/tests/fuzz_robustness.rs
@@ -314,7 +314,7 @@ fn mutate(rng: &mut XorShift, seed: &[u8]) -> String {
         match rng.below(3) {
             0 => {
                 let i = rng.below(bytes.len());
-                bytes[i] = (0x20 + rng.below(0x5f) as u8) as u8;
+                bytes[i] = 0x20 + rng.below(0x5f) as u8;
             }
             1 => {
                 let i = rng.below(bytes.len() + 1);
